+++ pytest_django/__init__.py
@@ -1,6 +1,6 @@
 from pytest_django.plugin import *
 from pytest_django.funcargs import *
-from pytest_django.marks import *
+from pytest_django.deprecated import *
 
 # When Python 2.5 support is dropped, these imports can be used instead:
 # from .plugin import *