+++ pytest_django/funcargs.py
@@ -1,19 +1,29 @@
 import copy
 
+import pytest
 
-from django.conf import settings
-from django.contrib.auth.models import User
-
-from django.test.client import RequestFactory, Client
-
-from .live_server_helper import (HAS_LIVE_SERVER_SUPPORT, LiveServer,
+from .django_compat import setup_databases
+from .lazy_django import skip_if_no_django
+from .live_server_helper import (has_live_server_support, LiveServer,
                                  get_live_server_host_ports)
+
+
+def pytest_funcarg__djangodb(request):
+    """Ensure the Django test database is loaded"""
+    # Not sure if there's much point in the marking
+    skip_if_no_django()
+    if not hasattr(request.function, 'djangodb'):
+        request.function.djangodb = pytest.mark.djangodb
+    setup_databases(request._pyfuncitem.session)
 
 
 def pytest_funcarg__client(request):
     """
     Returns a Django test client instance.
     """
+    skip_if_no_django()
+
+    from django.test.client import Client
     return Client()
 
 
@@ -21,6 +31,10 @@ def pytest_funcarg__admin_client(request):
     """
     Returns a Django test client logged in as an admin user.
     """
+    request.getfuncargvalue('djangodb')
+
+    from django.contrib.auth.models import User
+    from django.test.client import Client
 
     try:
         User.objects.get(username='admin')
@@ -41,6 +55,10 @@ def pytest_funcarg__rf(request):
     """
     Returns a RequestFactory instance.
     """
+    skip_if_no_django()
+
+    from django.test.client import RequestFactory
+
     return RequestFactory()
 
 
@@ -49,6 +67,10 @@ def pytest_funcarg__settings(request):
     Returns a Django settings object that restores any changes after the test
     has been run.
     """
+    skip_if_no_django()
+
+    from django.conf import settings
+
     old_settings = copy.deepcopy(settings)
 
     def restore_settings():
@@ -60,8 +82,13 @@ def pytest_funcarg__settings(request):
 
 
 def pytest_funcarg__live_server(request):
-    if not HAS_LIVE_SERVER_SUPPORT:
-        raise Exception('The liveserver funcarg is not supported in Django <= 1.3')
+    skip_if_no_django()
+
+    if not hasattr(request.function, 'djangodb'):
+        request.function.djangodb = pytest.mark.djangodb(transaction=True)
+
+    if not has_live_server_support():
+        pytest.fail('live_server tests is not supported in Django <= 1.3')
 
     def setup_live_server():
         return LiveServer(*get_live_server_host_ports())
@@ -69,4 +96,6 @@ def pytest_funcarg__live_server(request):
     def teardown_live_server(live_server):
         live_server.thread.join()
 
-    return request.cached_setup(setup=setup_live_server, teardown=teardown_live_server, scope='session')
+    return request.cached_setup(setup=setup_live_server,
+                                teardown=teardown_live_server,
+                                scope='session')