+++ pytest_django/marks.py
@@ -1,3 +1,0 @@
-import pytest
-
-transaction_test_case = pytest.mark.transaction_test_case