+++ pytest_django/plugin.py
@@ -7,62 +7,64 @@ test. Additionally, the settings are copied before each test and restored at
 the end of the test, so it is safe to modify settings within tests.
 """
 
-import sys
+import os
 
-from django.conf import settings
-from django.core.management import call_command
-from django.core import management
-from django.core.urlresolvers import clear_url_caches
-from django.test.simple import DjangoTestSuiteRunner
-
-import django.db.backends.util
-
-
-from .utils import is_django_unittest, django_setup_item, django_teardown_item
 from .db_reuse import monkey_patch_creation_for_db_reuse
+from .django_compat import (setup_databases, teardown_databases,
+                            is_django_unittest, clear_django_outbox,
+                            django_setup_item, django_teardown_item)
+from .lazy_django import django_settings_is_configured, skip_if_no_django
 
-import py
+import pytest
 
 
-suite_runner = None
-old_db_config = None
+def create_django_runner(reuse_db, create_db):
+    """Setup the Django test environment
 
+    Return an instance of DjangoTestSuiteRunner which can be used to
+    setup and teardown a Django test environment.
 
-def get_runner(config):
+    If ``reuse_db`` is True, if found, an existing test database will be used.
+    When no test database exists, it will be created.
+
+    If ``create_db`` is True, the test database will be created or re-created,
+    no matter what ``reuse_db`` is.
+    """
+
+    from django.test.simple import DjangoTestSuiteRunner
+
     runner = DjangoTestSuiteRunner(interactive=False)
-
-    if config.option.no_db:
-        def cursor_wrapper_exception(*args, **kwargs):
-            raise RuntimeError('No database access is allowed since --no-db was used!')
-
-        def setup_databases():
-            # Monkey patch CursorWrapper to warn against database usage
-            django.db.backends.util.CursorWrapper = cursor_wrapper_exception
-
-        def teardown_databases(db_config):
-            pass
-
-        runner.setup_databases = setup_databases
-        runner.teardown_databases = teardown_databases
-
-    elif config.option.reuse_db:
-
-        if not config.option.create_db:
+    if reuse_db:
+        if not create_db:
             monkey_patch_creation_for_db_reuse()
-
-        # Leave the database for the next test run
         runner.teardown_databases = lambda db_config: None
-
     return runner
 
 
+def get_django_settings_module(config):
+    """
+    Returns the value of DJANGO_SETTINGS_MODULE. The first specified value from
+    the following will be used:
+     * --ds command line option
+     * DJANGO_SETTINGS_MODULE pytest.ini option
+     * DJANGO_SETTINGS_MODULE
+
+    """
+    ordered_settings = [
+        config.option.ds,
+        config.getini('DJANGO_SETTINGS_MODULE'),
+        os.environ.get('DJANGO_SETTINGS_MODULE')
+    ]
+
+    for ds in ordered_settings:
+        if ds:
+            return ds
+
+    return None
+
+
 def pytest_addoption(parser):
-    group = parser.getgroup('django database setup')
-    group._addoption('--no-db',
-                     action='store_true', dest='no_db', default=False,
-                     help='Run tests without setting up database access. Any '
-                          'communication with databases will result in an '
-                          'exception.')
+    group = parser.getgroup('django')
 
     group._addoption('--reuse-db',
                      action='store_true', dest='reuse_db', default=False,
@@ -75,91 +77,166 @@ def pytest_addoption(parser):
                      help='Re-create the database, even if it exists. This '
                           'option will be ignored if not --reuse-db is given.')
 
+    group._addoption('--ds',
+                     action='store', type='string', dest='ds', default=None,
+                     help='Set DJANGO_SETTINGS_MODULE.')
 
-def _disable_south_management_command():
-    management.get_commands()
-    # make sure `south` migrations are disabled
-    management._commands['syncdb'] = 'django.core'
+    parser.addini('DJANGO_SETTINGS_MODULE',
+                  'Django settings module to use by pytest-django')
+
+
+def pytest_configure(config):
+    """Configure DJANGO_SETTINGS_MODULE and register our marks
+
+    The first specified value from the following will be used:
+
+    * --ds command line option
+    * DJANGO_SETTINGS_MODULE pytest.ini option
+    * DJANGO_SETTINGS_MODULE
+
+    It will set the "ds" config option regardless of the method used
+    to set DJANGO_SETTINGS_MODULE, allowing to check for the plugin
+    being used by doing `config.getvalue('ds')`.
+    """
+
+    ds = get_django_settings_module(config)
+
+    if ds:
+        os.environ['DJANGO_SETTINGS_MODULE'] = ds
+    else:
+        os.environ.pop('DJANGO_SETTINGS_MODULE', None)
+
+    # Register the marks
+    config.addinivalue_line(
+        'markers',
+        'djangodb(transaction=False, multidb=False): Mark the test as using '
+        'the django test database.  The *transaction* argument marks will '
+        "allow you to use transactions in the test like Django's "
+        'TransactionTestCase while the *multidb* argument will work like '
+        "Django's multi_db option on a TestCase: all test databases will be "
+        'flushed instead of just the default.')
+    config.addinivalue_line(
+        'markers',
+        'urls(modstr): Use a different URLconf for this test, similar to '
+        'the `urls` attribute of Django `TestCase` objects.  *modstr* is '
+        'a string specifying the module of a URL config, e.g. '
+        '"my_app.test_urls".')
 
 
 def pytest_sessionstart(session):
-    global suite_runner, old_db_config
+    if django_settings_is_configured():
+        # This import fiddling is needed to give a proper error message
+        # when the Django settings module cannot be found
+        try:
+            import django
+            django  # Silence pyflakes
+        except ImportError:
+            raise pytest.UsageError('django could not be imported, make sure '
+                                    'it is installed and available on your'
+                                    'PYTHONPATH')
 
-    _disable_south_management_command()
+        from django.conf import settings
 
-    suite_runner = get_runner(session.config)
-    suite_runner.setup_test_environment()
+        try:
+            # Make sure the settings actually gets loaded
+            settings.DATABASES
+        except ImportError, e:
+            # An import error here means that DJANGO_SETTINGS_MODULE could not
+            # be imported
+            raise pytest.UsageError(*e.args)
 
-    old_db_config = suite_runner.setup_databases()
+        runner = create_django_runner(
+            create_db=session.config.getvalue('create_db'),
+            reuse_db=session.config.getvalue('reuse_db'))
 
-    settings.DEBUG_PROPAGATE_EXCEPTIONS = True
+        runner.setup_test_environment()
+        settings.DEBUG_PROPAGATE_EXCEPTIONS = True
+        session.django_runner = runner
 
 
 def pytest_sessionfinish(session, exitstatus):
-    global suite_runner, old_db_config
-
-    capture = py.io.StdCapture()
-    suite_runner.teardown_test_environment()
-
-    suite_runner.teardown_databases(old_db_config)
-
-    unused_out, err = capture.reset()
-    sys.stderr.write(err)
+    runner = getattr(session.config, 'django_runner', None)
+    if runner:
+        teardown_databases(session)
+        runner.teardown_test_environment()
 
 
-_old_urlconf = None
+def validate_djangodb(marker):
+    """This function validates the djangodb marker
+
+    It checks the signature and creates the `transaction` and
+    `mutlidb` attributes on the marker which will have the correct
+    value.
+    """
+    # Use a fake function to check the signature
+    def apifun(transaction=False, multidb=False):
+        return transaction, multidb
+    marker.transaction, marker.multidb = apifun(*marker.args, **marker.kwargs)
 
 
-# trylast is needed to have access to funcargs
-@py.test.mark.trylast
+def validate_urls(marker):
+    """This function validates the urls marker
+
+    It checks the signature and creates the `urls` attribute on the
+    marker which will have the correct value.
+    """
+    def apifun(urls):
+        return urls
+    marker.urls = apifun(*marker.args, **marker.kwargs)
+
+
+# Trylast is needed to have access to funcargs, live_server suport
+# needs this and some funcargs add the djangodb marker which also
+# needs this to be called afterwards.
+@pytest.mark.trylast
 def pytest_runtest_setup(item):
-    global _old_urlconf
+    # Empty the django test outbox
+    if django_settings_is_configured():
+        clear_django_outbox()
 
     # Set the URLs if the pytest.urls() decorator has been applied
-    if hasattr(item.obj, 'urls'):
-        _old_urlconf = settings.ROOT_URLCONF
-        settings.ROOT_URLCONF = item.obj.urls
+    marker = getattr(item.obj, 'urls', None)
+    if marker:
+        skip_if_no_django()
+        from django.conf import settings
+        from django.core.urlresolvers import clear_url_caches
+
+        validate_urls(marker)
+        item.django_urlconf = settings.ROOT_URLCONF
+        settings.ROOT_URLCONF = marker.urls
         clear_url_caches()
 
-    # Invoke Django code to prepare the environment for the test run
-    if not item.config.option.no_db and not is_django_unittest(item):
+    if hasattr(item.obj, 'djangodb'):
+        # Setup Django databases
+        skip_if_no_django()
+        validate_djangodb(item.obj.djangodb)
+        setup_databases(item.session)
         django_setup_item(item)
+    elif django_settings_is_configured() and not is_django_unittest(item):
+        # Block access to the Django databases
+        import django.db.backends.util
+
+        def cursor_wrapper(*args, **kwargs):
+            __tracebackhide__ = True
+            __tracebackhide__  # Silence pyflakes
+            pytest.fail('Database access not allowed, '
+                        'use the "djangodb" mark to enable')
+
+        item.django_cursor_wrapper = django.db.backends.util.CursorWrapper
+        django.db.backends.util.CursorWrapper = cursor_wrapper
 
 
 def pytest_runtest_teardown(item):
-    global _old_urlconf
-
     # Call Django code to tear down
-    if not item.config.option.no_db and not is_django_unittest(item):
+    if django_settings_is_configured():
         django_teardown_item(item)
 
-    if hasattr(item, 'urls') and _old_urlconf is not None:
-        settings.ROOT_URLCONF = _old_urlconf
-        _old_urlconf = None
+    if hasattr(item, 'django_urlconf'):
+        from django.conf import settings
+        from django.core.urlresolvers import clear_url_caches
+        settings.ROOT_URLCONF = item.django_urlconf
         clear_url_caches()
 
-
-def pytest_namespace():
-    def load_fixture(fixture):
-        raise Exception('load_fixture is deprecated. Use a standard Django '
-                        'test case or invoke call_command("loaddata", fixture)'
-                        'instead.')
-
-    def urls(urlconf):
-        """
-        A decorator to change the URLconf for a particular test, similar
-        to the `urls` attribute on Django's `TestCase`.
-
-        Example:
-
-            @pytest.urls('myapp.test_urls')
-            def test_something(client):
-                assert 'Success!' in client.get('/some_path/')
-        """
-        def wrapper(function):
-            function.urls = urlconf
-            return function
-
-        return wrapper
-
-    return {'load_fixture': load_fixture, 'urls': urls}
+    if hasattr(item, 'django_cursor_wrapper'):
+        import django.db.backends.util
+        django.db.backends.util.CursorWrapper = item.django_cursor_wrapper