+++ tests/test_funcargs.py
@@ -1,3 +1,4 @@
+import pytest
 from django.test.client import Client, RequestFactory
 
 pytest_plugins = ['pytester']
@@ -7,6 +8,7 @@ def test_client(client):
     assert isinstance(client, Client)
 
 
+@pytest.mark.djangodb
 def test_admin_client(admin_client):
     assert isinstance(admin_client, Client)
     assert admin_client.get('/admin-required/').content == 'You are an admin'